breaking up into multiple projects
- `core` -> `docker-it-scala-core`
  + has all the docker services

- `scalatest` -> `docker-testkit-scalatest`
  + `DockerTestKit` for `scalatest`
  + all tests that were originally in this repository

- `specs2` -> `docker-testkit-specs2`
  + `DockerTestKit` for `specs2`
  + all tests originall in this repo, but adapted to `specs2`

The `DockerTestKit` used to be in package `whisk.docker.test`, but with
this new multibuild, the package for both `scalatest` and `specs2` is
`whisk.docker`.
