+++ core/src/main/scala/whisk/docker/Docker.scala
