+++ core/src/main/scala/whisk/docker/DockerCassandraService.scala
