+++ core/src/main/scala/whisk/docker/DockerContainer.scala
