+++ core/src/main/scala/whisk/docker/DockerContainerOps.scala
