+++ core/src/main/scala/whisk/docker/DockerElasticsearchService.scala
