+++ core/src/main/scala/whisk/docker/DockerKafkaService.scala
