+++ core/src/main/scala/whisk/docker/DockerKit.scala
