+++ core/src/main/scala/whisk/docker/DockerMongodbService.scala
