+++ core/src/main/scala/whisk/docker/DockerNeo4jService.scala
