+++ core/src/main/scala/whisk/docker/DockerReadyChecker.scala
