+++ scalatest/src/test/scala/whisk/docker/DockerServiceSpec.scala
