+++ scalatest/src/test/scala/whisk/docker/ElasticsearchServiceSpec.scala
@@ -1,0 +1,16 @@
+package whisk.docker
+
+import org.scalatest._
+import org.scalatest.concurrent.ScalaFutures
+import org.scalatest.time._
+
+class ElasticsearchServiceSpec extends FlatSpec with Matchers with BeforeAndAfterAll with GivenWhenThen with ScalaFutures
+    with DockerElasticsearchService with DockerTestKit {
+
+  implicit val pc = PatienceConfig(Span(20, Seconds), Span(1, Second))
+
+  "elasticsearch container" should "be ready" in {
+    elasticsearchContainer.isReady().futureValue shouldBe true
+  }
+
+}
