+++ specs2/src/main/scala/whisk/docker/BeforeAfterAllStopOnError.scala
@@ -1,0 +1,15 @@
+package whisk.docker
+
+import org.specs2.specification.core.SpecificationStructure
+import org.specs2.specification.create.FragmentsFactory
+import org.specs2.specification.core.Fragments
+
+trait BeforeAfterAllStopOnError extends SpecificationStructure
+    with FragmentsFactory {
+  def beforeAll
+  def afterAll
+  override def map(fs: => Fragments) =
+    super.map(fs).prepend(
+      fragmentFactory.step(beforeAll).stopOnError
+    ).append(fragmentFactory.step(afterAll))
+}
