--- src/test/scala/whisk/docker/AllAtOnceSpec.scala
@@ -1,17 +1,0 @@
-package whisk.docker
-
-import org.scalatest.concurrent.ScalaFutures
-import org.scalatest.time.{ Second, Seconds, Span }
-import org.scalatest.{ GivenWhenThen, BeforeAndAfterAll, Matchers, FlatSpec }
-import whisk.docker.test.DockerTestKit
-
-class AllAtOnceSpec extends FlatSpec with Matchers with BeforeAndAfterAll with GivenWhenThen with ScalaFutures
-    with DockerElasticsearchService with DockerCassandraService with DockerNeo4jService with DockerMongodbService with PingContainerKit {
-
-  implicit val pc = PatienceConfig(Span(20, Seconds), Span(1, Second))
-
-  "all containers" should "be ready at the same time" in {
-    dockerContainers.map(_.image).foreach(println)
-    dockerContainers.forall(_.isReady().futureValue) shouldBe true
-  }
-}
